+++ app/cart/page.tsx
@@ -4,6 +4,9 @@ import MenuOne from '@/components/Header/Menu/MenuOne'
 import Breadcrumb from '@/components/Breadcrumb/Breadcrumb'
 import Footer from '@/components/Footer/Footer'
 import CartClient from '@/components/Cart/CartClient'
+import { Metadata } from 'next'
+
+
 
 
 
@@ -16,10 +19,22 @@ const Cart = async () => {
                 <MenuOne props="bg-transparent" />
                 <Breadcrumb heading='Shopping cart' subHeading='Shopping cart' />
             </div>
-            <CartClient  />
+            <CartClient />
             <Footer />
         </>
     )
+}
+
+export const metadata: Metadata = {
+    title: 'Cart - SakibBaba Store',
+    description: 'View and manage your shopping cart at SakibBaba Store. Adjust quantities, remove items, and proceed to checkout.',
+    keywords: ['cart', 'shopping cart', 'SakibBaba Store', 'online store', 'checkout'],
+    openGraph: {
+        title: 'Cart - SakibBaba Store',
+        description: 'View and manage your shopping cart at SakibBaba Store. Adjust quantities, remove items, and proceed to checkout.',
+        type: 'website',
+        url: '/cart',
+    },
 }
 
 export default Cart
