+++ app/layout.tsx
@@ -4,15 +4,17 @@ import ModalQuickview from '@/components/Modal/ModalQuickview'
 import ModalSearch from '@/components/Modal/ModalSearch'
 import ModalWishlist from '@/components/Modal/ModalWishlist'
 import '@/styles/styles.scss'
-import type { Metadata } from 'next'
 import { Instrument_Sans } from 'next/font/google'
 import GlobalProvider from './GlobalProvider'
+import { Metadata } from 'next'
+
 
 const instrument = Instrument_Sans({ subsets: ['latin'] })
 
+
 export const metadata: Metadata = {
-  title: 'Anvogue',
-  description: 'Multipurpose eCommerce Template',
+  title: 'SakibBaba Store',
+  description: 'Your online store built with Next.js and WooCommerce',
 }
 
 export default function RootLayout({
